requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==1.10.15
email-validator==1.3.1
python-dotenv
requests
anyio
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
prometheus-client
psutil
